small toolbar changes
